Separate collapsed quick take into it's own component
